requirements.txt
streamlit
python-docx
PyPDF2
faiss-cpu
google-genai   
tqdm
numpy
scikit-learn
sentencepiece      